fix: поправлена выпадающая ошибка при изменении имени категории на существующую

+++ ewm-main-service/src/main/java/ru/practicum/explorewithme/service/category/service/CategoryServiceImpl.java
@@ -11,6 +11,8 @@ import ru.practicum.explorewithme.service.category.dto.NewCategoryRequest;
 import ru.practicum.explorewithme.service.category.dto.UpdateCategoryRequest;
 import ru.practicum.explorewithme.service.category.mapper.CategoryMapper;
 import ru.practicum.explorewithme.service.category.model.Category;
+import ru.practicum.explorewithme.service.event.dal.EventRepository;
+import ru.practicum.explorewithme.service.exception.ConflictException;
 import ru.practicum.explorewithme.service.exception.DuplicatedDataException;
 import ru.practicum.explorewithme.service.exception.NotFoundException;
 
@@ -21,11 +23,12 @@ import java.util.List;
 @Slf4j
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
     @Override
     public CategoryDto createCategory(NewCategoryRequest request) {
         if (categoryRepository.existsByName(request.getName())) {
-            throw new DuplicatedDataException("Категория " +
+            throw new ConflictException("Категория " +
                     request.getName() + " уже существует");
         }
         Category category = CategoryMapper.mapToCategory(request);
@@ -35,14 +38,17 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public CategoryDto changeCategory(Long catId, UpdateCategoryRequest request) {
+        if (!categoryRepository.findByNameIgnoreCaseAndIdNot(request.getName(), catId).isEmpty()) {throw new ConflictException("Такая категория уже существует" + request.getName());}
         Category category = categoryRepository.findById(catId).orElseThrow(() -> new NotFoundException("Категория " + catId + " не была найдена"));
         Category updatedCategory = CategoryMapper.mapToUpdateCategory(request, category);
+
         categoryRepository.save(updatedCategory);
         return CategoryMapper.mapToCategoryDto(updatedCategory);
     }
 
     @Override
     public void removeCategory(Long catId) {
+        if (eventRepository.countByCategoryId(catId) > 0) {throw new ConflictException("В этой категории уже существуют события");}
         Category category = categoryRepository.findById(catId).orElseThrow(() -> new NotFoundException("Категория " + catId + " не была найдена"));
         categoryRepository.delete(category);
     }
